[kpuzzle] Implement transformation application.

+++ src/kpuzzle/definition.rs
@@ -3,11 +3,12 @@ use std::{collections::HashMap, rc::Rc};
 use super::{state::KStateData, transformation::KTransformationData};
 
 // use super::super::{state::KStateData, transformation::KTransformationData};
-
+#[derive(Debug)]
 pub struct KPuzzleOrbitDefinition {
-    pub num_pieces: u32,
-    pub num_orientations: u32,
+    pub num_pieces: usize,       // TODO
+    pub num_orientations: usize, // TODO
 }
+#[derive(Debug)]
 pub struct KPuzzleDefinition {
     pub name: String,
     pub orbits: HashMap<String, KPuzzleOrbitDefinition>,
